+++ core/apis.go
@@ -300,4 +300,16 @@ func (s *systemImpl) SysDeleteApplicationTypeVersion(appTypeID string, versionID
 	return s.app.sysDeleteApplicationTypeVersion(appTypeID, versionID, l)
 }
 
+func (s *systemImpl) SysCreateAuthTypes(code string, description string, isExternal bool, isAnonymous bool, useCredentials bool, ignoreMFA bool, params map[string]interface{}) (*model.AuthType, error) {
+	return s.app.sysCreateAuthTypes(code, description, isExternal, isAnonymous, useCredentials, ignoreMFA, params)
+}
+
+func (s *systemImpl) SysGetAuthTypes() ([]model.AuthType, error) {
+	return s.app.sysGetAuthTypes()
+}
+
+func (s *systemImpl) SysUpdateAuthTypes(ID string, code string, description string, isExternal bool, isAnonymous bool, useCredentials bool, ignoreMFA bool, params map[string]interface{}) error {
+	return s.app.SysUpdateAuthTypes(ID, code, description, isExternal, isAnonymous, useCredentials, ignoreMFA, params)
+}
+
 ///