+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Expose application type versions system APIs [#404](https://github.com/rokwire/core-building-block/issues/404)
+- Expose System APIs for auth types [#362](https://github.com/rokwire/core-building-block/issues/362)
 - Expose grant roles to account Admin API [#383](https://github.com/rokwire/core-building-block/issues/383)
 - Expose grant permissions to account Admin API [#382](https://github.com/rokwire/core-building-block/issues/382)
 - Expose Admin API which gives an application account devices [#359](https://github.com/rokwire/core-building-block/issues/359)
@@ -19,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add app config endpoints [#261](https://github.com/rokwire/core-building-block/issues/261)
 
 ### Fixed
+- Fix issues with account linking [#408](https://github.com/rokwire/core-building-block/issues/408)
 - Fix creating application group admin API [#397](https://github.com/rokwire/core-building-block/issues/397)
 - Fix create role and create group Admin APIs [#386](https://github.com/rokwire/core-building-block/issues/386)
 - Fix broken app config API and CHANGELOG.md [#401](https://github.com/rokwire/core-building-block/issues/401)